remove the ssh2-python remote runner
The ssh2-python lib was introduced as an alternative to Paramiko, but it doesn't support sudo commands, which makes it usable only in a subset of cases.
Since the gains are marginal compared to other improvements that were made, we can remove it to simplify the code base and only retain with Paramiko.